requirements.txt
pandas
pytest
streamlit
google-genai
chromadb[default]
